pyproject.toml
[project]
name = "phd-lab-scraper"
version = "0.1.0"
description = "Collect ETHZ and EPFL labs and website links"
requires-python = ">=3.11"
dependencies = [
  "beautifulsoup4>=4.12.0",
  "pyyaml>=6.0",
  "requests>=2.31.0",
]

[dependency-groups]
dev = [
  "ruff>=0.13.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

